requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
chromedriver-binary-auto==0.3.1
colorama==0.4.6
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
numpy==1.25.2
outcome==1.2.0
packaging==23.1
pluggy==1.3.0
pyarrow==13.0.0
pycparser==2.21
pylance==0.6.5
PySocks==1.7.1
pytest==7.4.0
python-dotenv==1.0.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
wsproto==1.2.0
